refactor(scrollbar)!: move ScrollbarBehavior into scrolling/
All scrollbar config now lives under scrolling/ (ScrollbarStyle for
geometry, ScrollbarThemeData for colors, ScrollbarBehavior for
interaction). ScrollbarBehavior carries no design-system dependency, so
co-locating it with the widget was an asymmetry; the widget module now
holds only the widget.

BREAKING CHANGE: import path changes from
nuiitivet.widgets.scrollbar.ScrollbarBehavior to
nuiitivet.scrolling.ScrollbarBehavior.

Closes #264
