+++ src/nuiitivet/scrolling/__init__.py
@@ -2,6 +2,7 @@
 
 from .controller import ScrollAxisState, ScrollController
 from .scrollable_style import ScrollableStyle
+from .scrollbar_behavior import ScrollbarBehavior
 from .scrollbar_theme_data import ScrollbarThemeData
 from .scrollbar_style import ScrollbarStyle
 from .types import ScrollDirection, ScrollPhysics
@@ -10,6 +11,7 @@ __all__ = [
     "ScrollAxisState",
     "ScrollController",
     "ScrollableStyle",
+    "ScrollbarBehavior",
     "ScrollbarStyle",
     "ScrollbarThemeData",
     "ScrollDirection",
